+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/handler/BinaryExchangeHandler.java
@@ -17,11 +17,19 @@
 package de.gematik.test.tiger.proxy.handler;
 
 import de.gematik.rbellogger.data.RbelElement;
-import de.gematik.rbellogger.data.facet.*;
+import de.gematik.rbellogger.data.RbelElementConvertionPair;
+import de.gematik.rbellogger.data.RbelHostname;
+import de.gematik.rbellogger.data.facet.RbelBinaryFacet;
+import de.gematik.rbellogger.data.facet.RbelFacet;
+import de.gematik.rbellogger.data.facet.RbelMessageTimingFacet;
+import de.gematik.rbellogger.data.facet.RbelTcpIpMessageFacet;
+import de.gematik.rbellogger.data.facet.TigerNonPairedMessageFacet;
+import de.gematik.rbellogger.data.facet.TracingMessagePairFacet;
 import de.gematik.test.tiger.common.data.config.tigerproxy.DirectReverseProxyInfo;
 import de.gematik.test.tiger.mockserver.model.BinaryMessage;
-import de.gematik.test.tiger.mockserver.model.BinaryProxyListener;
 import de.gematik.test.tiger.proxy.TigerProxy;
+import de.gematik.test.tiger.proxy.exceptions.TigerProxyRoutingException;
+import de.gematik.test.tiger.proxy.exceptions.TigerRoutingErrorFacet;
 import java.net.SocketAddress;
 import java.net.SocketException;
 import java.time.ZoneId;
@@ -31,10 +39,11 @@ import java.util.Queue;
 import java.util.concurrent.CompletableFuture;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
+import lombok.val;
 
 @Data
 @Slf4j
-public class BinaryExchangeHandler implements BinaryProxyListener {
+public class BinaryExchangeHandler {
 
   private final TigerProxy tigerProxy;
   private final BinaryChunksBuffer binaryChunksBuffer;
@@ -46,7 +55,6 @@ public class BinaryExchangeHandler implements BinaryProxyListener {
             tigerProxy.getRbelLogger().getRbelConverter(), tigerProxy.getTigerProxyConfiguration());
   }
 
-  @Override
   public void onProxy(
       BinaryMessage binaryRequest,
       Optional<CompletableFuture<BinaryMessage>> binaryResponseFuture,
@@ -101,18 +109,16 @@ public class BinaryExchangeHandler implements BinaryProxyListener {
                         }
                       })
                   .exceptionally(
-                      t -> {
-                        DirectReverseProxyInfo reverseProxyInfo = Optional.ofNullable(getTigerProxy()
-                                        .getTigerProxyConfiguration()
-                                        .getDirectReverseProxy())
-                                .orElse(null);
-
-                        if (reverseProxyInfo == null || !reverseProxyInfo.isIgnoreConnectionErrors()) {
-                          if (isConnectionResetException(t)) {
-                            log.trace("Connection reset:", t);
+                      exception -> {
+                        if (!shouldIgnoreConnectionErrors()) {
+                          if (isConnectionResetException(exception)) {
+                            log.trace("Connection reset:", exception);
                           } else {
-                            log.warn("Exception during Direct-Proxy handling:", t);
-                            propagateExceptionMessageSafe(t);
+                            log.warn("Exception during Direct-Proxy handling:", exception);
+                            propagateExceptionMessageSafe(
+                                exception,
+                                RbelHostname.create(clientAddress),
+                                RbelHostname.create(serverAddress));
                           }
                         }
                         return null;
@@ -120,9 +126,16 @@ public class BinaryExchangeHandler implements BinaryProxyListener {
       log.trace("Returning from BinaryExchangeHandler!");
     } catch (RuntimeException e) {
       log.warn("Uncaught exception during handling of request", e);
-      propagateExceptionMessageSafe(e);
+      propagateExceptionMessageSafe(
+          e, RbelHostname.create(clientAddress), RbelHostname.create(serverAddress));
       throw e;
     }
+  }
+
+  private boolean shouldIgnoreConnectionErrors() {
+    return Optional.ofNullable(getTigerProxy().getTigerProxyConfiguration().getDirectReverseProxy())
+        .map(DirectReverseProxyInfo::isIgnoreConnectionErrors)
+        .orElse(false);
   }
 
   private static boolean isConnectionResetException(Throwable t) {
@@ -165,9 +178,29 @@ public class BinaryExchangeHandler implements BinaryProxyListener {
     return rbelMessageOptional;
   }
 
-  private void propagateExceptionMessageSafe(Throwable exception) {
+  public void propagateExceptionMessageSafe(
+      Throwable exception, RbelHostname senderAddress, RbelHostname receiverAddress) {
     try {
       tigerProxy.propagateException(exception);
+
+      final TigerProxyRoutingException routingException =
+          new TigerProxyRoutingException(
+              "Exception during handling of HTTP request: " + exception.getMessage(),
+              senderAddress,
+              receiverAddress,
+              exception);
+      log.info(routingException.getMessage(), routingException);
+
+      val message = new RbelElement(new byte[] {}, null);
+      message.addFacet(new TigerRoutingErrorFacet(routingException));
+      tigerProxy
+          .getRbelLogger()
+          .getRbelConverter()
+          .parseMessage(
+              new RbelElementConvertionPair(message),
+              senderAddress,
+              receiverAddress,
+              Optional.of(routingException.getTimestamp()));
     } catch (Exception handlingException) {
       log.warn(
           "While propagating an exception another error occured (ignoring):", handlingException);