+++ tiger-rbel/src/main/java/de/gematik/rbellogger/converter/RbelConverter.java
@@ -28,7 +28,6 @@ import de.gematik.rbellogger.util.RbelMessagesDequeFacade;
 import de.gematik.test.tiger.common.config.TigerTypedConfigurationKey;
 import de.gematik.test.tiger.common.util.TigerSecurityProviderInitialiser;
 import java.nio.charset.StandardCharsets;
-import java.text.MessageFormat;
 import java.time.ZonedDateTime;
 import java.util.*;
 import java.util.concurrent.CompletableFuture;
@@ -39,7 +38,6 @@ import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 import lombok.*;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.apache.commons.lang3.tuple.Pair;
 
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
@@ -112,25 +110,9 @@ public class RbelConverter {
       try {
         plugin.consumeElement(convertedInput, this);
       } catch (RuntimeException e) {
-        final String msg =
-            MessageFormat.format(
-                "Exception during conversion with plugin '{0}' ({1})",
-                plugin.getClass().getName(), e.getMessage());
-        log.info(msg);
-        log.debug("Stack trace", e);
-        if (log.isDebugEnabled()) {
-          log.debug(
-              "Content in failed conversion-attempt was (B64-encoded) {}",
-              Base64.getEncoder().encodeToString(convertedInput.getRawContent()));
-          if (convertedInput.getParentNode() != null) {
-            log.debug(
-                "Parent-Content in failed conversion-attempt was (B64-encoded) {}",
-                Base64.getEncoder().encodeToString(convertedInput.getParentNode().getRawContent()));
-          }
-        }
-        convertedInput.addFacet(
-            new RbelNoteFacet(
-                msg + "\n" + ExceptionUtils.getStackTrace(e), RbelNoteFacet.NoteStyling.ERROR));
+        val conversionException = RbelConversionException.wrapIfNotAConversionException(e, plugin, convertedInput);
+        conversionException.printDetailsToLog(log);
+        conversionException.addErrorNoteFacetToElement();
       }
     }
     return convertedInput;
@@ -423,9 +405,9 @@ public class RbelConverter {
         future.getKey().get();
       } catch (InterruptedException e) {
         Thread.currentThread().interrupt();
-        throw new RbelConversionException(e);
+        throw new RbelConversionException(e, future.getValue());
       } catch (ExecutionException e) {
-        throw new RbelConversionException(e);
+        throw new RbelConversionException(e, future.getValue());
       }
     }
   }