+++ tiger-rbel/src/main/java/de/gematik/rbellogger/data/facet/RbelMessageInfoFacet.java
@@ -1,4 +1,5 @@
 /*
+ *
  * Copyright 2024 gematik GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,16 +14,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package de.gematik.rbellogger.data.facet;
 
+import lombok.AllArgsConstructor;
+import lombok.Data;
 
-import lombok.Builder;
+/** Facet which carries information about how to represent this message when rendering */
+@Data
+@AllArgsConstructor
+public class RbelMessageInfoFacet implements RbelFacet {
+  private final String symbol;
+  private final String color;
+  private final String menuInfoString;
+  private final String abbrev;
+  private final String title;
 
-public class RbelResponseFacet extends RbelMessageInfoFacet {
-
-  @Builder
-  public RbelResponseFacet(String menuInfoString) {
-    super("fa-reply", "text-success", menuInfoString, "RES", "Response");
+  public static RbelFacet newErrorSymbol(String menuInfoString) {
+    return new RbelMessageInfoFacet("fa-triangle-exclamation", "text-danger", menuInfoString, "ERR", "Error");
   }
 }